Move RouteForm types to a separate file
- RouteForm's types are used in a lot of places. Move them to a separate file and rename them to make their usage easier, more readable and to avoid dependency cycles.
