+++ src/components/map/routes/Routes.tsx
@@ -1,6 +1,9 @@
 import React, { useContext } from 'react';
 import { MapEditorContext, Mode } from '../../../context/MapEditor';
-import { FormState, schema } from '../../forms/RoutePropertiesForm';
+import {
+  routeFormSchema,
+  RouteFormState,
+} from '../../forms/RoutePropertiesForm.types';
 import { CreateRouteModal } from './CreateRouteModal';
 
 export const Routes: React.FC = () => {
@@ -16,7 +19,7 @@ export const Routes: React.FC = () => {
   const routeDetails = editedRouteData.metaData;
   // checking whether 'routeDetails' already contains all the information necessary
   // if not -> should show the form
-  const areFormValuesValid = schema.safeParse(routeDetails).success;
+  const areFormValuesValid = routeFormSchema.safeParse(routeDetails).success;
   const showCreateForm = !areFormValuesValid && drawingMode === Mode.Draw;
 
   const onCancel = () => {
@@ -26,7 +29,7 @@ export const Routes: React.FC = () => {
     dispatch({ type: 'stopDrawRoute' });
   };
 
-  const onSuccess = (data: FormState) => {
+  const onSuccess = (data: RouteFormState) => {
     dispatch({
       type: 'setState',
       payload: {
@@ -38,7 +41,7 @@ export const Routes: React.FC = () => {
     });
   };
 
-  const defaultValues: Partial<FormState> = routeDetails || {};
+  const defaultValues: Partial<RouteFormState> = routeDetails || {};
 
   return (
     <>
